+++ SweatBoxTrainer/aircraft.h
@@ -133,7 +133,7 @@ public:
 	bool arrived();
 	AssignedValues& getAssignedValues() { return assignedValues; }
 	double calculateGS(double __unnamed000, double __unnamed001, double gs_angle, double dest_altitude);
-	double calculateLoc(double __unnamed000, double __unnamed001, int __unnamed002);
+	double calculateLoc(double __unnamed000, double __unnamed001, double __unnamed002);
 	bool calculateTurnDistance(Point2* p, Point2* p2);
 	bool defaultTurnDistance();
 	Point2* getFuturePoint();
